userservice/requirements.txt
Flask>=2.0.2
gunicorn
psycopg2-binary
pandas
openpyxl
prometheus-flask-exporter